tests: add custom lib name integration test
This test copies the `test-crate` integration test, but customizes the
`lib` target's name in `test-workspace/test-crate-custom-lib-name/Cargo.toml`
to ensure that the tool works as expected even when the lib target uses
a name other than the crate name. The expected output is updated to
reflect the change in crate name, and the custom lib name.